Separated AIStates and UtilityAI classes into individual classes. No more inheritance. Saving work.
