+++ Assets/Scripts/CombatAIState.cs
@@ -14,29 +14,45 @@ namespace ScavengerWorld.AI
 
     public class CombatAIState : AIState
     {
+        private CombatUtilityAI ai;
+
         public bool IsBlocking { get; set; }
 
-        public CombatAIState(NavMeshAgent navigator, Unit unit, AnimationController animController)
-            : base(navigator, unit, animController)
+        public bool Enabled { get; set; }
+
+        public override bool CanEnterState => Enabled;
+
+        public override bool CanExitState => !Enabled;
+
+        public CombatAIState(Unit unit, NavMeshAgent navigator, AnimationController animController)
         {
-            this.navigator = navigator;
             this.unit = unit;
+            this.mover = new(unit, navigator);
             this.animController = animController;
+
             this.locomotion = unit.Weapon?.combatLocomotion;
             this.ai = new CombatUtilityAI();
-            this.state = State.Combat;
+            this.mode = AIMode.Combat;
             actionProgress = 0;
-        }
-
-        public void SetTarget(Interactable target)
-        {
-            ai.Target = target;
-        }
+        }        
 
         public override void OnEnterState()
         {
-            navigator.speed = 8f;
+            mover.SetMaxSpeed(8f);
             AnimateLocomotion();
+            ai.SelectTarget(unit);
+            //ai.Target.Unit.AIController.SetState(State.Combat, this.unit.Interactable);
+        }
+
+        public override void OnExitState()
+        {
+            ai.Reset();
+        }
+
+        public void AddTarget(Interactable target)
+        {
+            ai.AddTarget(target);
+            //ai.Target = target;
         }
 
         public override void OnUpdate()
@@ -48,46 +64,27 @@ namespace ScavengerWorld.AI
             // Will need to constantly check for an update in enemy targets or
             // listen to some event fired when enemy engages or is nearby
 
-            locomotion.State.Parameter = navigator.velocity.magnitude;
-            HandleRotation();
+            locomotion.State.Parameter = mover.Speed;
+            mover.HandleRotation();
 
             if (selectedAction is null)
             {
-
-
-                // Decide which enemy target to focus
-
-                // Select attack move
                 selectedAction = ai.DecideBestAction(unit);
                 selectedAction.Target = ai.Target;
                 return;
             }
 
-            if (HasReachedTarget())
+            if (mover.HasReachedTarget(selectedAction.Target))
             {
                 if (!selectedAction.IsRunning)
                 {
-                    DisableMovement();
+                    mover.DisableMovement();
                     selectedAction.StartAction();
                 }
             }
             else
             {
-                MoveToTarget();
-            }
-        }
-
-        public override bool HasReachedTarget()
-        {
-            float distance = Vector3.Distance(unit.transform.position, ai.Target.transform.position);
-            return distance <= ai.Target.useRange;
-        }
-
-        private void MoveToTarget()
-        {
-            if (MoveEnabled)
-            {
-                navigator.SetDestination(ai.Target.transform.position);
+                mover.MoveToTarget(selectedAction.Target);
             }
         }
     }
