+++ Assets/Scripts/Mover.cs
@@ -1,158 +1,78 @@
-using System;
+using System.Collections;
 using System.Collections.Generic;
-using System.Xml;
-using Unity.MLAgents.Actuators;
 using UnityEngine;
 using UnityEngine.AI;
-using UnityEngine.Events;
 
 namespace ScavengerWorld
 {
-    /// <summary>
-    /// Responsible for handling all movement related logic
-    /// </summary>
-    public class Mover : MonoBehaviour
-    {        
-        [Range(5f, 100f)]
-        [SerializeField] private float focusRange = 100f;
-        [Range(20f, 50f)]
-        [Tooltip("Max range from (0,0,0) to alternatively explore if AI told agent to go off the map")]
-        [SerializeField] private float explorationRange = 40f;
-        [Range(10f, 30f)]
-        [SerializeField] private float rotateSpeed = 20f;
-        [SerializeField] private LayerMask interactableLayer;
-        [SerializeField] private NavMeshAgent navigator;
-                
+    public class Mover
+    {
         private Unit unit;
-        private ActionRunner actionRunner;
+        private NavMeshAgent navigator;
         private Vector3 facing;
-        private InteractNode targetSpot;
 
-        public event UnityAction OnReachedDestination;
-
-        public Unit Unit => unit;
-        public float NavigatorSpeed => navigator.velocity.magnitude;
-        public float StopDistance => navigator.stoppingDistance;
+        private Vector3 targetPos;
 
         public bool MoveEnabled { get; private set; }
+        public float Speed => navigator.velocity.magnitude;
 
-        public InteractNode TargetSpot => targetSpot;
-
-        public Interactable TargetInteractable { get; set; }
-
-        private void Awake()
+        public Mover(Unit unit, NavMeshAgent navigator)
         {
-            unit = GetComponent<Unit>();
-            actionRunner = GetComponent<ActionRunner>();
-            navigator = GetComponent<NavMeshAgent>();
-            navigator.autoRepath = true;
-            navigator.updateRotation = false;
-            
-            MoveEnabled = true;
-            TargetInteractable = null;
+            this.unit = unit;
+            this.navigator = navigator;
+            facing = unit.transform.forward;
         }
 
-        void Update()
+        public void SetMaxSpeed(float speed)
         {
-            HandleRotation();
+            navigator.speed = speed;
+        }
 
-            if (TargetInteractable != null)
+        public bool HasReachedTargetNode(InteractNode node)
+        {
+            float distance = Vector3.Distance(unit.transform.position, node.transform.position);
+            return Mathf.Abs(distance - navigator.stoppingDistance) <= 0.1f;
+        }
+
+        public bool HasReachedTarget(Interactable target)
+        {
+            float distance = Vector3.Distance(unit.transform.position, target.transform.position);
+            return Mathf.Abs(distance - target.useRange) <= 0.1f;
+        }
+
+        public bool HasReachedTargetPos()
+        {
+            float distance = Vector3.Distance(unit.transform.position, targetPos);
+            return Mathf.Abs(distance - navigator.stoppingDistance) <= 0.1f;
+        }
+
+        public void ResetTargetPos()
+        {
+            targetPos = default;
+        }
+
+        public void MoveToTargetNode(InteractNode node)
+        {
+            if (MoveEnabled)
             {
-                if (targetSpot != null)
-                {
-                    if (HasReachedTargetSpot())
-                    {
-                        if (!actionRunner.ActionIsRunning)
-                        {
-                            StopMoving();
-                            OnReachedDestination?.Invoke();
-                        }                        
-                    }
-                    else
-                    {
-                        MoveToTargetSpot();
-                    }
-                }
-                else
-                {
-                    if (TargetInteractable.InteractionAvailable)
-                    {
-                        targetSpot = TargetInteractable.GetAvailableInteractNode(this.unit);
-                        if (targetSpot != null)
-                        {
-                            targetSpot.SetOccupant(unit);
-                        }
-                        else
-                        {
-                            actionRunner.CancelCurrentAction();
-                        }
-                    }
-                    else
-                    {
-                        actionRunner.CancelCurrentAction();
-                    }                    
-                }
+                navigator.SetDestination(node.transform.position);
             }
-
-            //if (TargetInteractable is null && TargetPos != default)
-            //{
-            //    if (HasReachedTargetPos())
-            //    {
-            //        StopMoving();
-            //        targetPos = default;
-            //        return;
-            //    }
-            //    return;
-            //}
-
-            //if (TargetInteractable != null && TargetPos == default)
-            //{
-            //    if (HasReachedTargetInteractable())
-            //    {
-            //        if (!actionRunner.ActionIsRunning)
-            //        {
-            //            StopMoving();
-            //            OnReachedInteractableTarget?.Invoke();
-            //            return;
-            //        }
-            //    }
-            //    else
-            //    {
-            //        //if (!actionRunner.ActionIsRunning)
-            //        //{
-            //        //    MoveToInteractable(TargetInteractable);
-            //        //}
-            //        MoveToInteractable(TargetInteractable);
-            //    }
-
-            //    // TODO: Need logic here to tell agent to recalculate new path if stuck
-            //}
         }
 
-        public bool HasReachedTargetSpot()
+        public void MoveToTarget(Interactable target)
         {
-            float distance = Vector3.Distance(transform.position, targetSpot.transform.position);
-            return Mathf.Abs(distance - navigator.stoppingDistance) <= 0.01;
+            if (MoveEnabled)
+            {
+                navigator.SetDestination(target.transform.position);
+            }
         }
 
-        /// <summary>
-        /// Use this to move character to clicked position
-        /// </summary>
-        /// <param name="pos"></param>
         public void MoveToPosition(Vector3 pos)
         {
             if (MoveEnabled)
             {
-                actionRunner.CancelCurrentAction();
+                targetPos = pos;
                 navigator.SetDestination(pos);
-            }            
-        }
-
-        public void MoveToTargetSpot()
-        {
-            if (MoveEnabled)
-            {
-                navigator.SetDestination(targetSpot.transform.position);
             }
         }
 
@@ -169,132 +89,36 @@ namespace ScavengerWorld
 
         public void DisableMovement()
         {
-            StopMoving();
+            navigator.velocity = Vector3.zero;
+            navigator.ResetPath();
             MoveEnabled = false;
+        }
+
+        public void FaceTowards(Interactable i)
+        {
+            facing = i.transform.position - unit.transform.position;
+            facing.y = 0f;
+            facing.Normalize();
+
+            //Apply Rotation
+            Quaternion targ_rot = Quaternion.LookRotation(facing, Vector3.up);
+            Quaternion nrot = Quaternion.RotateTowards(unit.transform.rotation, targ_rot, 360f);
+            unit.transform.rotation = nrot;
         }
 
         public void HandleRotation()
         {
             if (navigator.hasPath)
             {
-                facing = navigator.steeringTarget - transform.position;
+                facing = navigator.steeringTarget - unit.transform.position;
                 facing.y = 0f;
                 facing.Normalize();
 
                 //Apply Rotation
                 Quaternion targ_rot = Quaternion.LookRotation(facing, Vector3.up);
-                Quaternion nrot = Quaternion.Slerp(transform.rotation, targ_rot, rotateSpeed * Time.deltaTime);
-                transform.rotation = nrot;
+                Quaternion nrot = Quaternion.Slerp(unit.transform.rotation, targ_rot, unit.RotateSpeed * Time.deltaTime);
+                unit.transform.rotation = nrot;
             }
         }
-
-        public void FaceTowards(Interactable i)
-        {
-            facing = i.transform.position - transform.position;
-            facing.y = 0f;
-            facing.Normalize();
-
-            //Apply Rotation
-            Quaternion targ_rot = Quaternion.LookRotation(facing, Vector3.up);
-            Quaternion nrot = Quaternion.RotateTowards(transform.rotation, targ_rot, 360f);
-            transform.rotation = nrot;
-        }
-
-        public void ClearTarget()
-        {
-            targetSpot?.ClearOccupant();
-            targetSpot = null;
-            TargetInteractable = null;
-        }
-
-        #region Old Stuff
-
-        public bool FoodIsNearby(out Gatherable gatherable)
-        {
-            Collider[] colliders = Physics.OverlapSphere(transform.position, focusRange, interactableLayer);
-
-            if (colliders.Length == 0) 
-            {
-                gatherable = null;
-                return false;
-            };
-
-            foreach (Collider c in colliders)
-            {
-                gatherable = c.GetComponent<Gatherable>();
-                if (gatherable != null) return true;
-            }
-            gatherable=null;
-            return false;
-        }
-
-        public bool EnemyUnitIsNearby(out Unit enemyUnit)
-        {
-            Collider[] colliders = Physics.OverlapSphere(transform.position, focusRange, interactableLayer);
-
-            if (colliders.Length == 0)
-            {
-                enemyUnit = null;
-                return false;
-            }                
-
-            foreach (Collider c in colliders)
-            {
-                enemyUnit = c.GetComponent<Unit>();
-                if (enemyUnit != null
-                    && enemyUnit.TeamId != this.unit.TeamId
-                    && !enemyUnit.IsStorageDepot)
-                {
-                    return true;
-                }
-            }
-            enemyUnit=null;
-            return false;
-        }
-
-        public bool EnemyStorageIsNearby(out Unit enemyStorage)
-        {
-            Collider[] colliders = Physics.OverlapSphere(transform.position, focusRange, interactableLayer);
-
-            if (colliders.Length == 0)
-            {
-                enemyStorage = null;
-                return false;
-            }
-
-            foreach (Collider c in colliders)
-            {
-                enemyStorage = c.GetComponent<Unit>();
-                if (enemyStorage != null
-                    && enemyStorage.TeamId != this.unit.TeamId
-                    && enemyStorage.IsStorageDepot)
-                {
-                    return true;
-                }
-            }
-            enemyStorage=null;
-            return false;
-        }
-
-        public bool IsValidPath(Vector3 pos)
-        {
-            NavMeshPath path = new();
-            navigator.CalculatePath(pos, path);
-            return path.status == NavMeshPathStatus.PathComplete;
-        }
-
-        public Vector3 GetValidPath(Vector3 currentPos)
-        {
-            Vector3 randomPos = UnityEngine.Random.insideUnitSphere * 20f + Vector3.zero;
-            NavMesh.SamplePosition(randomPos, out NavMeshHit hit, explorationRange, NavMesh.AllAreas);
-            return hit.position;
-        }
-
-        public void ResetNavigator()
-        {
-            navigator.ResetPath();
-        }
-
-        #endregion
     }
 }
